krishimitra-ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]
python-dotenv==1.0.1
python-multipart==0.0.9

# Gemini LLM
google-generativeai==0.7.2

# Pinecone vector DB
pinecone-client==4.1.0

# Whisper speech-to-text
openai-whisper
torch==2.3.1
torchaudio==2.3.1

# Text-to-speech
gTTS==2.5.1

# Embeddings (converts text to vectors for Pinecone)
# paraphrase-multilingual-mpnet-base-v2 supports Marathi, Hindi, English
sentence-transformers==3.0.1

# PDF + Word doc parsing (for knowledge base ingestion)
pypdf2==3.0.1
python-docx==1.1.0

# MongoDB (for chat history)
motor==3.5.1

# Audio processing
pydub==0.25.1